extract parametric-editor
